requirements.txt
bleach==6.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
imbalanced-learn==0.12.3
imblearn==0.0
joblib==1.4.2
kaggle==1.6.17
lightgbm==4.5.0
numpy==2.0.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
text-unidecode==1.3
threadpoolctl==3.5.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
xgboost==2.1.0
